Remove old value serializer code
