+++ src/NexusMods.MnemonicDB.Abstractions/Attribute.cs
@@ -96,6 +96,27 @@ public abstract partial class Attribute<TValueType, TLowLevelType> : IAttribute<
         var prefix = datom.Prefix;
         return new ReadDatom(in prefix, ReadValue(datom.ValueSpan, datom.Prefix.ValueTag, resolver), this);
     }
+    
+    /// <summary>
+    /// Reads the high level value from the given span
+    /// </summary>
+    public TValueType ReadValue(ReadOnlySpan<byte> span, ValueTag tag, AttributeResolver resolver)
+    {
+        return FromLowLevel(tag.Read<TLowLevelType>(span), resolver);
+    }
+    
+    /// <summary>
+    /// Write a datom for this attribute to the given writer
+    /// </summary>
+    public void Write<TWriter>(EntityId entityId, AttributeCache cache, TValueType value, TxId txId, bool isRetract, TWriter writer)
+        where TWriter : IBufferWriter<byte>
+    {
+        var prefix = new KeyPrefix(entityId, cache.GetAttributeId(Id), txId, isRetract, LowLevelType);
+        var span = writer.GetSpan(KeyPrefix.Size);
+        MemoryMarshal.Write(span, prefix);
+        writer.Advance(KeyPrefix.Size);
+        LowLevelType.Write(ToLowLevel(value), writer);
+    }
 
     /// <summary>
     /// Returns true if the attribute is present on the entity
